requirements.txt
pytest==7.4.0
Faker==19.3.0
PyAutoGUI~=0.9.54
PyGetWindow~=0.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
allure-pytest==2.13.2
allure-python-commons==2.13.2